requirements.txt
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.0
ffmpeg-python==0.2.0
filelock==3.10.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
pycodestyle==2.10.0
pycparser==2.21
regex==2022.10.31
requests==2.28.2
semantic-version==2.10.0
setuptools-rust==1.5.2
sounddevice==0.4.6
soundfile==0.12.1
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
